codes/jackal_diffusion/requirements.txt
numpy
torch
matplotlib
hydra-core
omegaconf
dill
